applied same changes to deleteProject
